[refactor] Moving AWS files into their own package
