+++ src/main/java/com/amazon/jenkins/ec2fleet/Registry.java
@@ -1,11 +1,12 @@
 package com.amazon.jenkins.ec2fleet;
 
+import com.amazon.jenkins.ec2fleet.aws.CloudFormationApi;
+import com.amazon.jenkins.ec2fleet.aws.EC2Api;
+
 /**
  * Decouple plugin code from dependencies for easy testing. We cannot just make transient fields
  * in required classes as they usually restored by Jenkins without constructor call. Instead
  * we are using registry pattern.
- *
- * @see EC2FleetCloud
  */
 @SuppressWarnings("WeakerAccess")
 public class Registry {
