+++ src/test/java/com/amazon/jenkins/ec2fleet/aws/EC2ApiTest.java
@@ -1,5 +1,6 @@
-package com.amazon.jenkins.ec2fleet;
+package com.amazon.jenkins.ec2fleet.aws;
 
+import com.amazon.jenkins.ec2fleet.aws.EC2Api;
 import com.amazonaws.services.ec2.AmazonEC2;
 import com.amazonaws.services.ec2.model.AmazonEC2Exception;
 import com.amazonaws.services.ec2.model.CreateTagsRequest;
